requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-cors==0.8.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorful==0.5.6
cryptography==44.0.2
Deprecated==1.2.18
distlib==0.3.9
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.18.0
frozenlist==1.5.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
Jinja2==3.1.4
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lean-dojo==2.2.0
loguru==0.7.3
lxml==5.3.2
MarkupSafe==2.1.5
mpmath==1.3.0
msgpack==1.1.0
multidict==6.3.2
networkx==3.4.2
numpy==2.1.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==1.75.0
opencensus==0.11.4
opencensus-context==0.1.3
packaging==24.2
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.7
prometheus_client==0.21.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
py-spy==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
PyGithub==2.6.1
PyJWT==2.10.1
PyNaCl==1.5.0
python-dotenv==1.1.0
PyYAML==6.0.2
ray==2.44.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
rsa==4.9
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sympy==1.13.1
toml==0.10.2
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.67.1
triton==3.2.0
types-psutil==7.0.0.20250401
types-toml==0.10.8.20240310
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
virtualenv==20.30.0
wrapt==1.17.2
yarl==1.19.0
